Boxplot S4 Object Class

Objects of class Boxplot are used to store information for the Boxplot and Violin plots. Like all S4 objects in Slick, slots in this object can be accessed and assigned using functions corresponding to slot name. See Boxplot() and the the See Also section below.

Details

Objects of class Boxplot are created with Boxplot()

Multi-Language Support

Text with multi-language supported can be provided as a named list. Available languages:

en: English (default)

es: Spanish

fr: French

Note

Character strings in Code, Label, and Description must all be same length as the number of performance indicators (nPIs) in Value

Slots

Code

A short code for the Performance Indicators for this object. A character string length nPI or a named list for multi-language support. See Details

Label

A short label for the Performance Indicators for this object. Used to label axes on charts. Can be longer than Code but recommended to keep short as possible so it shows clearly in plots and tables. A character string length nPI or a named list for multi-language support. See Details

Description

A description for the Performance Indicators for this object. Can include Markdown, see Examples. A character string length nPI or a named list for multi-language support. See Details

Value

A numeric array with the stochastic performance indicator values for each simulation (sim), operating model (OM), management procedure (MP), and performance indicator (PI). Dimensions: c(nsim, nOM, nMP, and nPI).

Preset

An optional named list for the preset buttons in the App(). The name of the list element will appear as a button in the App().

Defaults

A list object with default selections for the Boxplot. See Boxplot()

See also

Examples

boxplot <- Boxplot(Code=c('PI1', 'PI2', 'PI3'),
                   Label=c('Performance Indicator 1',
                         'Performance Indicator 2',
                         'Performance Indicator 3'),
                   Description = c('This is the description for PI 1',
                                   'This is the description for PI 2',
                                   'This is the description for PI 3'),
                   Value=array(runif(3000), dim=c(6, 5, 4, 3)),
                   Preset=list('PI and P2'=1:2)

)

boxplot
#> 
#> ── An object of class `Boxplot` ────────────────────────────────────────────────
#> 
#> ── `Code` ──
#> 
#> 1 PI1
#> 2 PI2
#> 3 PI3
#> 
#> ── `Label` ──
#> 
#> 1 Performance Indicator 1
#> 2 Performance Indicator 2
#> 3 Performance Indicator 3
#> 
#> ── `Description` ──
#> 
#> 1 This is the description for PI 1
#> 2 This is the description for PI 2
#> 3 This is the description for PI 3
#> 
#> ── `Value` ──
#> 
#>   Dimension Length
#> 1      nsim      6
#> 2       nOM      5
#> 3       nMP      4
#> 4       nPI      3
#> 
#> ℹ Mean across MPs and PIs
#>       PI1  PI2  PI3
#> MP 1 0.46 0.54 0.52
#> MP 2 0.49 0.57 0.50
#> MP 3 0.45 0.49 0.49
#> MP 4 0.58 0.57 0.52
#> 
#> ── `Preset` ──
#> 
#> ── PI and P2 
#> 1
#> 2

Code(boxplot) <- c('LTY', 'STY', 'P50')
Metadata(boxplot)
#>   Code                   Label                      Description
#> 1  LTY Performance Indicator 1 This is the description for PI 1
#> 2  STY Performance Indicator 2 This is the description for PI 2
#> 3  P50 Performance Indicator 3 This is the description for PI 3

# Add to `Slick` object
slick <- Slick()
Boxplot(slick) <- boxplot

plotBoxplot(slick)
#> Error in data.frame(Sim = 1:nsim, OM = rep(OM_labels, each = nsim), MP = rep(MP_lab,     each = nsim * nOM), PI = rep(PI_names, each = nsim * nOM *     nMP), value = as.vector(values)): arguments imply differing number of rows: 6, 0, 30, 360

plotBoxplot(slick, type='violin')
#> Error in data.frame(Sim = 1:nsim, OM = rep(OM_labels, each = nsim), MP = rep(MP_lab,     each = nsim * nOM), PI = rep(PI_names, each = nsim * nOM *     nMP), value = as.vector(values)): arguments imply differing number of rows: 6, 0, 30, 360

plotBoxplot(slick, type='both')
#> Error in data.frame(Sim = 1:nsim, OM = rep(OM_labels, each = nsim), MP = rep(MP_lab,     each = nsim * nOM), PI = rep(PI_names, each = nsim * nOM *     nMP), value = as.vector(values)): arguments imply differing number of rows: 6, 0, 30, 360