S4 class Kobe

Slots can be accessed and assigned using functions corresponding to slot name. See See Also section below.

Details

Objects of class Kobe are created with Kobe()

Performance Indicators

By default, the first PI will be on the x-axis (usually B/BMSY or something similar) and the second on the y-axis (e.g., F/FMSY). The Slick App() provides dropdowns for selecting other PIs.

Multi-Language Support

Text with multi-language supported can be provided as a named list. Available languages:

en: English (default)

es: Spanish

fr: French

Note

Character strings in Code, Label, and Description must all be same length as the number of performance indicators (nPIs) in Value

Slots

Code

A short code for the Performance Indicators for this object. A character string length nPI or a named list for multi-language support. See Details

Label

A short label for the Performance Indicators for this object. Used to label axes on charts. Can be longer than Code but recommended to keep short as possible so it shows clearly in plots and tables. A character string length nPI or a named list for multi-language support. See Details

Description

A description for the Performance Indicators for this object. Can include Markdown, see Examples. A character string length nPI or a named list for multi-language support. See Details

Time

A numeric vector with values for the projection time-steps. Must match length nTS in Value

TimeLab

Character string length 1. Name of the time step (e.g., 'Year'). Will be used as the label in the Kobe Time plot. Use a named list for multiple languages.

Value

A numeric array with the stochastic performance indicator values for each simulation (sim), operating model (OM), management procedure (MP), performance indicator (PI), and projection time-steps (nTS) Dimensions: c(nsim, nOM, nMP, nPI, nTS)

Preset

An optional named list for the preset buttons in the App(). The name of the list element will appear as a button in the App().

Target

Numeric vector length nPI with the target value for the PIs. Defines the color quadrants on the Kobe plot. Defaults to 1.

Limit

Numeric vector length nPI with the limit value for the two PIs. Shows as red line on Kobe plot. NULL to ignore.

Defaults

A list object with default selections for the Kobe See Kobe-methods()

See also

Examples

kobe <- Kobe(Code=c('PI1', 'PI2', 'PI3'),
                   Label=c('Performance Indicator 1',
                           'Performance Indicator 2',
                           'Performance Indicator 3'),
                   Description = c('This is the description for PI 1',
                                   'This is the description for PI 2',
                                   'This is the description for PI 3'),
                   Value=array(runif(600), dim=c(10, 2, 2,3, 5)),
                   Preset=list('PI and P2'=1:2)

)

kobe
#> 
#> ── An object of class `Kobe` ───────────────────────────────────────────────────
#> 
#> ── `Code` ──
#> 
#> 1 PI1
#> 2 PI2
#> 3 PI3
#> 
#> ── `Label` ──
#> 
#> 1 Performance Indicator 1
#> 2 Performance Indicator 2
#> 3 Performance Indicator 3
#> 
#> ── `Description` ──
#> 
#> 1 This is the description for PI 1
#> 2 This is the description for PI 2
#> 3 This is the description for PI 3
#> 
#> ── `Time` ──
#> 
#> ── `TimeLab` ──
#> 
#> Year
#> 
#> ── `Value` ──
#> 
#>   Dimension Length
#> 1      nsim     10
#> 2       nOM      2
#> 3       nMP      2
#> 4       nPI      3
#> 5       nTS      5
#> 
#> ℹ Mean across MPs and PIs
#>       PI1  PI2  PI3
#> MP 1 0.49 0.49 0.51
#> MP 2 0.50 0.51 0.52
#> 
#> ── `Preset` ──
#> 
#> ── PI and P2 
#> 1
#> 2
#> 
#> ── `Target` ──
#> 
#> PI1 PI2 PI3 
#>   1   1   1 
#> ── `Limit` ──
#> 

# ADD TIME

# Code(boxplot)
# Metadata(boxplot)
#
# # Value(boxplot)
# Value(boxplot) <- array()
#
# Preset(boxplot)
# Preset(boxplot) <- list()
#
# boxplot