MPs S4 class and functions

An MPs object contains information about the management procedures (MPs) in the Slick() object.

Usage

MPs(x, ...)

# S4 method for dataframe_list
MPs(x, ...)

# S4 method for missingOrNULL
MPs(x, ...)

# S4 method for Slick
MPs(x, lang = NULL)

MPs(Slick) <- value

# S4 method for Slick
MPs(Slick) <- value

Arguments

x

An object of class data.frame() or list()

...

additional arguments (not used)

lang

optional.character string length 2 to specify language (if available). 'en', 'es', or 'fr'

Slick

An object of class Slick()

value

An object of class MPs()

missingOrNULL

missing or NULL

Details

Metadata

The Metadata data.frame must have n MPs rows and the following columns:

Code: A short character string for identifying the MPs. Used in the App() in places where there is no room for a full label.

Label: A long character string (one word or max two words) for identifying the MPs.

Description: A longer character string providing a description of the MPs.

Color: A character vector with the colors to use to display the MPs in the App(). Optional. If not provided, the App will try to set some sensible colors.

For multi-language support, the Metadata slot can be populated with a named list: names en, es, and fr for English, Spanish, and French respectively. The data.frame within each list element must have identical structure; i.e., the same number of rows and the column names in each list element must be identical.

The Metadata slot is accessed with Metadata(MPs) and modified with Metadata(MPs) <- data.frame()

Preset

The Preset slot is an optional named list to add preset buttons for the Management Procedure filters in the App(). The name of the list element will appear as a button in the App(). Each list element should contain numeric values specifying the MPs to include. The values must be <= n MPs.

The Preset slot is accessed with Preset(MPs) and modified with Preset(MPs) <- list()`

Methods (by generic)

MPs(dataframe_list): Create a new object of class MPs()

MPs(missingOrNULL): Create a new object of class MPs()

MPs(Slick): Return the MPs() object from an object of class Slick()

MPs(Slick) <- value: Assign an object of class MPs() to an object of class Slick()

Functions

MPs(): A generic function

MPs(Slick) <- value: Assign an object of class MPs() to an object of class Slick()

Slots

Metadata

A data.frame with a specific structure, describing the details of the management procedures. See Details section.

Preset

An optional named list of preset buttons for filters in the App(). See Details section.

Examples

myMPs <- MPs()
Metadata(myMPs) <- data.frame(Code=c('MP1', 'MP2', 'MP3'),
                              Label=c('Management Procedure 1',
                                      'Management Procedure 2',
                                      'Management Procedure 3'),
                              Description=c('This is the description for Management Procedure 1',
                                            'This is the description for Management Procedure 2',
                                            'This is the description for Management Procedure 3')
                              )
Preset(myMPs) <- list(All=1:3, FirstTwo=1:2)

# Multi-language
df_es <- Metadata(myMPs)
df_fr <- Metadata(myMPs)
df_es$Description <- c("This is the Spanish description for Management Procedure 1",
                       "This is the Spanish description for Management Procedure 2",
                       "This is the Spanish description for Management Procedure 3")
df_fr$Description <- c("This is the French description for Management Procedure 1",
                       "This is the French description for Management Procedure 2",
                       "This is the French description for Management Procedure 3")

Metadata(myMPs) <- list(en=Metadata(myMPs),
                        es=df_es,
                        fr=df_fr)

Metadata(myMPs)
#>   Code                  Label
#> 1  MP1 Management Procedure 1
#> 2  MP2 Management Procedure 2
#> 3  MP3 Management Procedure 3
#>                                          Description
#> 1 This is the description for Management Procedure 1
#> 2 This is the description for Management Procedure 2
#> 3 This is the description for Management Procedure 3
Metadata(myMPs, 'es')
#>   Code                  Label
#> 1  MP1 Management Procedure 1
#> 2  MP2 Management Procedure 2
#> 3  MP3 Management Procedure 3
#>                                                  Description
#> 1 This is the Spanish description for Management Procedure 1
#> 2 This is the Spanish description for Management Procedure 2
#> 3 This is the Spanish description for Management Procedure 3
Metadata(myMPs, 'fr')
#>   Code                  Label
#> 1  MP1 Management Procedure 1
#> 2  MP2 Management Procedure 2
#> 3  MP3 Management Procedure 3
#>                                                 Description
#> 1 This is the French description for Management Procedure 1
#> 2 This is the French description for Management Procedure 2
#> 3 This is the French description for Management Procedure 3