Return Code, Label, Description and other information from an object

Return Code, Label, Description and other information from an object

Usage

Metadata(object, lang = "en")

Metadata(object) <- value

# S4 method for class 'Boxplot'
Metadata(object, lang = "en")

# S4 method for class 'Boxplot'
Metadata(object) <- value

# S4 method for class 'Kobe'
Metadata(object, lang = "en")

# S4 method for class 'Kobe'
Metadata(object) <- value

# S4 method for class 'MPs'
Metadata(object, lang = "en")

# S4 method for class 'MPs'
Metadata(object) <- value

# S4 method for class 'Quilt'
Metadata(object, lang = "en")

# S4 method for class 'Quilt'
Metadata(object) <- value

# S4 method for class 'Tradeoff'
Metadata(object, lang = "en")

# S4 method for class 'Tradeoff'
Metadata(object) <- value

# S4 method for class 'Timeseries'
Metadata(object, lang = "en")

# S4 method for class 'Timeseries'
Metadata(object) <- value

# S4 method for class 'Spider'
Metadata(object, lang = "en")

# S4 method for class 'Spider'
Metadata(object) <- value

# S4 method for class 'Slick'
Metadata(object, lang = "en")

Arguments

object

A Slick-class(), MPs-class(), Boxplot-class(), Kobe-class(), Quilt-class(), Spider-class(), Timeseries-class(), or Tradeoff-class() object

lang

Optional text string specifying the language (if available). Either 'en', 'es', or 'fr' for English, Spanish, or French respectively

value

Replacement value for Metadata() in the corresponding object. See help documentation for the relevant object class for details.

Value

A data.frame

Methods (by class)

Metadata(Boxplot): Return Metadata for Boxplot-class() objects

Metadata(Boxplot) <- value: Assign Metadata for Boxplot-class() objects

Metadata(Kobe): Return Metadata for Kobe-class() objects

Metadata(Kobe) <- value: Assign Metadata for Kobe-class() objects

Metadata(MPs): Return Metadata for MPs-class() objects

Metadata(MPs) <- value: Assign Metadata for MPs-class() objects

Metadata(Quilt): Return Metadata for Quilt-class() objects

Metadata(Quilt) <- value: Assign Metadata for Quilt-class() objects

Metadata(Tradeoff): Return Metadata for Tradeoff-class() objects

Metadata(Tradeoff) <- value: Assign Metadata for Tradeoff-class() objects

Metadata(Timeseries): Return Metadata for Timeseries-class() objects

Metadata(Timeseries) <- value: Assign Metadata for Timeseries-class() objects

Metadata(Spider): Return Metadata for Spider-class() objects

Metadata(Spider) <- value: Assign Metadata for Spider-class() objects

Metadata(Slick): Return Author, Email, and Institution from Slick() objects