Methods for Creating, Accessing and Assigning Quilt objects

The Quilt function is used both to create and modify an Quilt-class() object. and to access and assign Quilt for an object of class Slick-class(). See Details.

Usage

Quilt(
  Code = "",
  Label = "",
  Description = "",
  Value = array(),
  Preset = list(),
  Color = c("darkblue", "lightblue"),
  MinValue = as.numeric(NA),
  MaxValue = as.numeric(NA),
  Misc = list()
)

Quilt(Slick) <- value

# S4 method for class 'missing'
Quilt()

# S4 method for class 'character_list'
Quilt(
  Code = "",
  Label = "",
  Description = "",
  Value = array(),
  Preset = list(),
  Color = c("darkblue", "lightblue"),
  MinValue = as.numeric(NA),
  MaxValue = as.numeric(NA),
  Misc = list()
)

# S4 method for class 'Slick'
Quilt(Code)

# S4 method for class 'Slick'
Quilt(Slick) <- value

Arguments

Code

A short code for the Performance Indicators for this object. A character string length nPI or a named list for multi-language support. See Details

Label

A short label for the Performance Indicators for this object. Used to label axes on charts. Can be longer than Code but recommended to keep short as possible so it shows clearly in plots and tables. A character string length nPI or a named list for multi-language support. See Details

Description

A description for the Performance Indicators for this object. Can include Markdown, see Examples. A character string length nPI or a named list for multi-language support. See Details

Value

A numeric array with the stochastic performance indicator values for each operating model (OM), management procedure (MP), and performance indicator (PI). Dimensions: c(nOM, nMP, and nPI). Alternatively, to calculate average over both simulations and Operating Models, Value can be a 4-dimensional array with dimensions: c(nSim, nOM, nMP, and nPI).

Preset

An optional named list for the preset buttons in the App(). The name of the list element will appear as a button in the App().

Color

A character vector length 2 of colors for the maximum and minimum values in the chart.

MinValue

Numeric vector length nPI with the minimum possible value for the respective PIs. Defaults to minimum PI value in Value (averaged across OMs in some cases)

MaxValue

Numeric vector length nPI with the maximum possible value (i.e., best performance) for the respective PIs. Defaults to maximum PI value in Value (averaged across OMs in some cases).

Misc

A named list for additional miscellaneous information.

Slick

A Slick-class() object

value

A Quilt-class() object

Details

Objects of class Quilt are created with Quilt()

Use the Code(), Label(), Description(), Value(), Preset(), Color(), MinValue(), and MaxValue() functions to access and assign the values for an existing Quilt object, see Examples

Multi-Language Support

Text with multi-language supported can be provided as a named list. Available languages:

en: English (default)

es: Spanish

fr: French

pt: Portuguese

Note

Character strings in Code, Label, and Description must all be same length as the number of performance indicators (nPIs) in Value

Functions

Quilt(missing): Create an empty Quilt object

Quilt(character_list): Create a populated Quilt object

Quilt(Slick): Return Quilt from a Slick-class() object

Quilt(Slick) <- value: Assign a Quilt-class() object to a Slick-class() object

See also

Examples

# Generate dummy values
nOM <- 2
nMP <- 4
nPI <- 4

values <- array(NA, dim=c(nOM, nMP, nPI))

pi_means <- runif(nPI, 5, 50)
for (om in 1:nOM) {
  for (mp in 1:nMP) {
    for (pi in 1:nPI) {
      values[om, mp, pi] <- rlnorm(1,log(pi_means[pi]), 0.4)
    }
  }
}

# Create and populate Object
quilt <- Quilt(Code=c('PI1', 'PI2', 'PI3', 'PI4'),
               Label=c('Performance Indicator 1',
                       'Performance Indicator 2',
                       'Performance Indicator 3',
                       'Performance Indicator 4'),
               Description = c('This is the description for PI 1',
                               'This is the description for PI 2',
                               'This is the description for PI 3',
                               'This is the description for PI 4'),
               Value=values)

# Check
Check(quilt)
#> 
#> ── Checking: "Quilt" ──
#> 
#> ✔ Complete

# Add to `Slick` object
slick <- Slick()
Quilt(slick) <- quilt

# Plots
plotQuilt(slick)
#> ℹ Note: `MPs` is empty. Using default MP names and colors

# Alternative - include Simulation dimension

# Generate dummy values
nSim <- 3
nOM <- 2
nMP <- 4
nPI <- 4

values <- array(NA, dim=c(nSim, nOM, nMP, nPI))

pi_means <- runif(nPI, 5, 50)
for (om in 1:nOM) {
  for (mp in 1:nMP) {
    for (pi in 1:nPI) {
      values[, om, mp, pi] <- rlnorm(nSim,log(pi_means[pi]), 0.4)
    }
  }
}

# Create and populate Object
quilt <- Quilt(Code=c('PI1', 'PI2', 'PI3', 'PI4'),
               Label=c('Performance Indicator 1',
                       'Performance Indicator 2',
                       'Performance Indicator 3',
                       'Performance Indicator 4'),
               Description = c('This is the description for PI 1',
                               'This is the description for PI 2',
                               'This is the description for PI 3',
                               'This is the description for PI 4'),
               Value=values)

# Add to `Slick` object
slick <- Slick()
Quilt(slick) <- quilt

# Plots
plotQuilt(slick)
#> ℹ Note: `MPs` is empty. Using default MP names and colors

apply(quilt@Value, 3:4, mean) |> round(1)
#>      [,1] [,2] [,3] [,4]
#> [1,] 45.4 50.2 45.8 52.2
#> [2,] 51.0 56.9 33.8 58.9
#> [3,] 32.6 49.6 52.0 60.1
#> [4,] 37.2 44.1 78.7 56.7