Create a Slick object

Each slot of the Slick object can be accessed or modified by the corresponding accessor function. See Examples sections below for more details.

Usage

Slick()

Arguments

Title

Title for the Slick object. A character string. For multiple languages, use a named list with names: en, es, fr for the three supported languages.

Subtitle

Subtitle for the Slick object. A character string or a named list with languages: en, es, fr

Date

Either an object of class Date or character string. Date the Slick object was created. Defaults to Sys.Date()

Author

A character vector with name(s) of author(s) for the Slick object

Email

A character vector with email addresses for the author(s). Optional. Must be same length as Author

Institution

A character vector with institution details for the author(s). Optional. Must be same length as Author

Introduction

Introduction text for the Slick object. Supports all markdown formatting. For multiple languages, use a named list with names: en, es, fr for the three supported languages.

MPs

An object of class MPs()

OMs

An object of class OMs()

Boxplot

An object of class Boxplot()

Kobe

An object of class Kobe()

Quilt

An object of class Quilt()

Spider

An object of class Spider()

Timeseries

An object of class Timeseries()

Tradeoff

An object of class Tradeoff()

Examples

mySlick <- Slick()
Title(mySlick) <- 'This is the title'
Subtitle(mySlick) <- 'This is the subtitle'
Date(mySlick) <- Sys.Date()

Title(mySlick)
#> [1] "This is the title"
Subtitle(mySlick)
#> [1] "This is the subtitle"
Date(mySlick)
#> [1] "2024-05-29"