Create a Slick class object

The Slick class is the main object class used in the Slick package. It contains sub-objects for the management procedures MPs(), operating models OMs(), and the six chart types: Boxplot(), Kobe(), Quilt(), Spider(), Timeseries(), and Tradeoff(), as well as metadata information for the Slick object such as Title, Author, and Introduction.

Usage

Slick(
  Title = "",
  Subtitle = "",
  Date = Sys.Date(),
  Author = "",
  Email = "",
  Institution = "",
  Introduction = "",
  MPs = NULL,
  OMs = NULL,
  Boxplot = NULL,
  Kobe = NULL,
  Quilt = NULL,
  Spider = NULL,
  Timeseries = NULL,
  Tradeoff = NULL
)

Arguments

Title

Title for the Slick object. A character string. For multiple languages, use a named list with names: en, es, fr for the three supported languages.

Subtitle

Subtitle for the Slick object. A character string or a named list with languages: en, es, fr

Date

Date the Slick object was created. Text in format 'YYYY-MM-DD' or class Date e.g., Sys.Date()

Author

A character vector with Author(s) names. The length of the vector should equal the number of authors.

Email

A character vector with email addresses for the author(s). Must be same length as Author. Can include Markdown.

Institution

A character vector with institution details for the author(s). Must be same length as Author. Can include Markdown.

Introduction

Introduction text for the Slick object. Supports all markdown formatting. Character string, must be length 1. For multiple languages, use a named list with names: en, es, fr for the three supported languages.

MPs

An object of class MPs-class()

OMs

An object of class OMs-class()

Boxplot

An object of class Boxplot-class()

Kobe

An object of class Kobe-class()

Quilt

An object of class Quilt-class()

Spider

An object of class Spider-class()

Timeseries

An object of class Timeseries-class()

Tradeoff

An object of class Tradeoff-class()

Details

Objects of class Slick are created with Slick().

Like all S4 objects in Slick, slots in this object can be accessed and assigned using functions corresponding to slot name. See Usage and Functions section.

Multi-Language Support

Text with multi-language supported can be provided as a named list. Available languages:

en: English (default)

es: Spanish

fr: French

All functions with the exception of Date support Markdown.

Functions

Slick(): Create a Slick-class() object

Slots

Title

Title for the Slick object. A character string. For multiple languages, use a named list with names: en, es, fr for the three supported languages.

Subtitle

Subtitle for the Slick object. A character string or a named list with languages: en, es, fr

Date

Date the Slick object was created. Text in format 'YYYY-MM-DD' or class Date e.g., Sys.Date()

Author

A character vector with Author(s) names. The length of the vector should equal the number of authors.

Email

A character vector with email addresses for the author(s). Must be same length as Author. Can include Markdown.

Institution

A character vector with institution details for the author(s). Must be same length as Author. Can include Markdown.

Introduction

Introduction text for the Slick object. Supports all markdown formatting. Character string, must be length 1. For multiple languages, use a named list with names: en, es, fr for the three supported languages.

MPs

An object of class MPs-class()

OMs

An object of class OMs-class()

Boxplot

An object of class Boxplot-class()

Kobe

An object of class Kobe-class()

Quilt

An object of class Quilt-class()

Spider

An object of class Spider-class()

Timeseries

An object of class Timeseries-class()

Tradeoff

An object of class Tradeoff-class()

See also

Examples

library(Slick)
# Assign values to a new `Slick` object
slick <- Slick()

Title(slick) <- 'An Example Slick Object'
Subtitle(slick) <- "[Created for the Developer's Guide](https://slick.bluematterscience.com/articles/DevelopersGuide.html)"
Date(slick) <- Sys.Date()
Author(slick) <- 'Adrian Hordyk'
Email(slick) <-  "[mailto:adrian@bluematterscience.com](mailto:adrian@bluematterscience.com)"
Institution(slick) <- "[Blue Matter Science](bluematterscience.com)"

Introduction(slick) <- "
This is an example Slick object. It has been created for the purposes of demonstrating the key features of Slick.

The introduction can include paragraphs as well as Markdown such as *italics* and **bold** text and [Links](https://slick.bluematterscience.com).
"

# Access values from `Slick` object
Title(slick)
#> [1] "An Example Slick Object"
Subtitle(slick)
#> [1] "[Created for the Developer's Guide](https://slick.bluematterscience.com/articles/DevelopersGuide.html)"
Date(slick)
#> [1] "2024-07-01"
Author(slick)
#> [1] "Adrian Hordyk"
Email(slick)
#> [1] "[mailto:adrian@bluematterscience.com](mailto:adrian@bluematterscience.com)"
Institution(slick)
#> [1] "[Blue Matter Science](bluematterscience.com)"
Introduction(slick)
#> [1] "\nThis is an example Slick object. It has been created for the purposes of demonstrating the key features of Slick.\n\nThe introduction can include paragraphs as well as Markdown such as *italics* and **bold** text and [Links](https://slick.bluematterscience.com).\n"

# See ?MPs, ?OMs, ?Boxplot, ?Quilt, etc for examples on populating other
# objects