Access or assign Time for object of class Kobe or Timeseries

Access or assign Time for object of class Kobe or Timeseries

Usage

Time(object)

Time(object) <- value

# S4 method for class 'Kobe'
Time(object)

# S4 method for class 'Kobe'
Time(object) <- value

# S4 method for class 'Timeseries'
Time(object)

# S4 method for class 'Timeseries'
Time(object) <- value

Arguments

object

A Kobe-class() or Timeseries() class object

value

Value to assign to Time

Methods (by class)

Time(Kobe): Return Time from a Kobe-class() object

Time(Kobe) <- value: Assign Time to a Kobe-class() object

Time(Timeseries): Return Time from a Timeseries-class() object

Time(Timeseries) <- value: Assign Time to a Timeseries-class() object

Functions

Time(object) <- value: Assign value to object@Time