Methods for Creating, Accessing and Assigning Tradeoff objects

The Tradeoff function is used both to create and modify an Tradeoff-class() object. and to access and assign Tradeoff for an object of class Slick-class(). See Details.

Usage

Tradeoff(
  Code = "",
  Label = "",
  Description = "",
  Value = array(),
  Preset = list()
)

Tradeoff(Slick) <- value

# S4 method for missing
Tradeoff()

# S4 method for character
Tradeoff(
  Code = "",
  Label = "",
  Description = "",
  Value = array(),
  Preset = list()
)

# S4 method for list
Tradeoff(
  Code = "",
  Label = "",
  Description = "",
  Value = array(),
  Preset = list()
)

# S4 method for Slick
Tradeoff(Code)

# S4 method for Slick
Tradeoff(Slick) <- value

Arguments

Code

A short code for the Performance Indicators for this object. A character string length nPI or a named list for multi-language support. See Details

Label

A short label for the Performance Indicators for this object. Used to label axes on charts. Can be longer than Code but recommended to keep short as possible so it shows clearly in plots and tables. A character string length nPI or a named list for multi-language support. See Details

Description

A description for the Performance Indicators for this object. Can include Markdown, see Examples. A character string length nPI or a named list for multi-language support. See Details

Value

A numeric array with the stochastic performance indicator values for each operating model (OM), management procedure (MP), performance indicator (PI), and historical + projection timestep (nTS). Dimensions: c(nOM, nMP, nPI`)

Preset

An optional named list for the preset buttons in the App(). The name of the list element will appear as a button in the App().

Slick

A Slick-class() object

value

A Tradeoff-class() object

Details

Objects of class Tradeoff are created with Tradeoff()

Multi-Language Support

Text with multi-language supported can be provided as a named list. Available languages:

en: English (default)

es: Spanish

fr: French

Note

Character strings in Code, Label, and Description must all be same length as the number of performance indicators (nPIs) in Value

Functions

Tradeoff(missing): Create an empty Tradeoff object

Tradeoff(character): Create a populated Tradeoff object

Tradeoff(list): Create a populated Tradeoff object

Tradeoff(Slick): Return Tradeoff from a Slick-class() object

Tradeoff(Slick) <- value: Assign a Tradeoff-class() object to a Slick-class() object

See also