Plot Timeseries

Plots the historical and projected values for a performance indicator.

Usage

plotTimeseries(
  slick,
  PM_ind = 1,
  MP_ind = NULL,
  OM_ind = NULL,
  includeHist = TRUE,
  col_line = "darkgray",
  fill_ribbon1 = "#ededed",
  col_ribbon1 = "#ededed",
  quants1 = c(0.25, 0.75),
  fill_ribbon2 = "white",
  col_ribbon2 = "#c9c9c9",
  linetype_ribbon2 = "dashed",
  quants2 = c(0.1, 0.9),
  MP_label = "Code",
  size.title = 18,
  size.axis.title = 14,
  size.axis.text = 14,
  size.mp.label = 6,
  targ_color = "green",
  targ_name = "Target",
  lim_color = "red",
  lim_name = "Limit",
  inc_y_label = TRUE,
  sims = NULL
)

Arguments

slick

A Slick-class() object

PM_ind

A numeric value specifying the performance indicator to plot

MP_ind

A numeric value specifying the MP to plot. Defaults to NULL which shows all MPs

OM_ind

A numeric value specifying the OM to plot. Defaults to NULL where values are calculated as median across OMs.

includeHist

Logical. Include the historical period in the projections?

col_line

Color for the median line (historical)

fill_ribbon1

Fill color for the inner ribbon

col_ribbon1

Color of the line for inner ribbon

quants1

Quantiles for the inner ribbon. Numeric length 2

fill_ribbon2

Fill color for the outer ribbon

col_ribbon2

Color of the line for outer ribbon

linetype_ribbon2

Line type for outer ribbon

quants2

Quantiles for the outer ribbon. Numeric length 2.

MP_label

Label to use for the MPs. Either Code or Label. Description works as well, but you probably don't want to do that.

size.title

Numeric length 1. Size for plot title

size.axis.title

Numeric length 1. Size for axis title

size.axis.text

Numeric length 1. Size for axis text

size.mp.label

Numeric length 1. Size of MP labels. Set to NULL for no MP labels

targ_color

Color for the target line (if it exists in Target(Timeseries(slick)))

targ_name

Label for the target line

lim_color

Color for the limit line (if it exists in Limit(Timeseries(slick)))

lim_name

Label for the limit line

inc_y_label

Include the label for the y-axis?

Value

A ggplot2 object

Details

By default the Time Series chart plots the median, 25th and 75th quantiles as a shaded ribbon, and the 10th and 90th quantiles as dashed lines.

If the plot is by management procedure (MP), the quantiles will be included in the projections. If multiple MPs are plotted together, only the median across simulations will be shown in the projections.

If OM_ind isn't specified, the results will be shown as the median across operating models.

See also