Access or assign Code, Label, and Description for a valid object class

Access or assign Code, Label, and Description for a valid object class

Usage

Code(object, lang = "en")

Code(object) <- value

Description(object, lang = "en")

Description(object) <- value

Label(object, lang = "en")

Label(object) <- value

# S4 method for class 'Boxplot'
Code(object, lang = "en")

# S4 method for class 'Boxplot'
Code(object) <- value

# S4 method for class 'Boxplot'
Description(object, lang = "en")

# S4 method for class 'Boxplot'
Label(object, lang = "en")

# S4 method for class 'Boxplot'
Label(object) <- value

# S4 method for class 'Boxplot'
Description(object) <- value

# S4 method for class 'Kobe'
Code(object, lang = "en")

# S4 method for class 'Kobe'
Code(object) <- value

# S4 method for class 'Kobe'
Description(object, lang = "en")

# S4 method for class 'Kobe'
Description(object) <- value

# S4 method for class 'Kobe'
Label(object, lang = "en")

# S4 method for class 'Kobe'
Label(object) <- value

# S4 method for class 'MPs'
Code(object, lang = "en")

# S4 method for class 'MPs'
Code(object) <- value

# S4 method for class 'MPs'
Label(object, lang = "en")

# S4 method for class 'MPs'
Label(object) <- value

# S4 method for class 'MPs'
Description(object, lang = "en")

# S4 method for class 'MPs'
Description(object) <- value

# S4 method for class 'Quilt'
Code(object, lang = "en")

# S4 method for class 'Quilt'
Code(object) <- value

# S4 method for class 'Quilt'
Description(object, lang = "en")

# S4 method for class 'Quilt'
Description(object) <- value

# S4 method for class 'Quilt'
Label(object, lang = "en")

# S4 method for class 'Quilt'
Label(object) <- value

# S4 method for class 'Tradeoff'
Code(object, lang = "en")

# S4 method for class 'Tradeoff'
Code(object) <- value

# S4 method for class 'Tradeoff'
Description(object, lang = "en")

# S4 method for class 'Tradeoff'
Description(object) <- value

# S4 method for class 'Tradeoff'
Label(object, lang = "en")

# S4 method for class 'Tradeoff'
Label(object) <- value

# S4 method for class 'Timeseries'
Code(object, lang = "en")

# S4 method for class 'Timeseries'
Code(object) <- value

# S4 method for class 'Timeseries'
Description(object, lang = "en")

# S4 method for class 'Timeseries'
Description(object) <- value

# S4 method for class 'Timeseries'
Label(object, lang = "en")

# S4 method for class 'Timeseries'
Label(object) <- value

# S4 method for class 'Spider'
Code(object, lang = "en")

# S4 method for class 'Spider'
Code(object) <- value

# S4 method for class 'Spider'
Description(object, lang = "en")

# S4 method for class 'Spider'
Description(object) <- value

# S4 method for class 'Spider'
Label(object, lang = "en")

# S4 method for class 'Spider'
Label(object) <- value

Arguments

object

An object of class MPs-class(), Boxplot-class(), Kobe-class(), Quilt-class(), Spider-class(), Timeseries-class(), or Tradeoff-class()

lang

Optional text string specifying the language (if available). Either 'en', 'es', or 'fr' for English, Spanish, or French respectively

value

A character vector or a named list for multi-language support

Details

Code, Label, and Description must all be equal length.

Multi-Language Support

Text with multi-language supported can be provided as a named list. Available languages:

en: English (default)

es: Spanish

fr: French

Methods (by class)

Code(Boxplot): Return Code from a Boxplot-class() object

Code(Boxplot) <- value: Assign Code to a Boxplot-class() object

Description(Boxplot): Return Description from a Boxplot-class() object

Label(Boxplot): Return Label from a Boxplot-class() object

Label(Boxplot) <- value: Assign Label to a Boxplot-class() object

Description(Boxplot) <- value: Assign Description to a Boxplot-class() object

Code(Kobe): Return Code from a Kobe-class() object

Code(Kobe) <- value: Assign Code to a Kobe-class() object

Description(Kobe): Return Description from a Kobe-class() object

Description(Kobe) <- value: Assign Description to a Kobe-class() object

Label(Kobe): Return Label from a Kobe-class() object

Label(Kobe) <- value: Assign Label to a Kobe-class() object

Code(MPs): Return Code from a MPs-class() object

Code(MPs) <- value: Assign Code to a MPs-class() object

Label(MPs): Return Label from a MPs-class() object

Label(MPs) <- value: Assign Label to a MPs-class() object

Description(MPs): Return Description from a MPs-class() object

Description(MPs) <- value: Assign Description to a MPs-class() object

Code(Quilt): Return Code from a Quilt-class() object

Code(Quilt) <- value: Assign Code to a Quilt-class() object

Description(Quilt): Return Description from a Quilt-class() object

Description(Quilt) <- value: Assign Description to a Quilt-class() object

Label(Quilt): Return Label from a Quilt-class() object

Label(Quilt) <- value: Assign Label to a Quilt-class() object

Code(Tradeoff): Return Code from a Tradeoff-class() object

Code(Tradeoff) <- value: Assign Code to a Tradeoff-class() object

Description(Tradeoff): Return Description from a Tradeoff-class() object

Description(Tradeoff) <- value: Assign Description to a Tradeoff-class() object

Label(Tradeoff): Return Label from a Tradeoff-class() object

Label(Tradeoff) <- value: Assign Label to a Tradeoff-class() object

Code(Timeseries): Return Code from a Timeseries-class() object

Code(Timeseries) <- value: Assign Code to a Timeseries-class() object

Description(Timeseries): Return Description from a Timeseries-class() object

Description(Timeseries) <- value: Assign Description to a Timeseries-class() object

Label(Timeseries): Return Label from a Timeseries-class() object

Label(Timeseries) <- value: Assign Label to a Timeseries-class() object

Code(Spider): Return Code from a Spider-class() object

Code(Spider) <- value: Assign Code to a Spider-class() object

Description(Spider): Return Description from a Spider-class() object

Description(Spider) <- value: Assign Description to a Spider-class() object

Label(Spider): Return Label from a Spider-class() object

Label(Spider) <- value: Assign Label to a Spider-class() object

See also