S4 class Quilt

Objects of class Quilt are used to store information for the Quilt chart. Like all S4 objects in Slick, slots in this object can be accessed and assigned using functions corresponding to slot name. See Quilt() and the the See Also section below.

Details

Objects of class Quilt are created with Quilt()

Multi-Language Support

Text with multi-language supported can be provided as a named list. Available languages:

en: English (default)

es: Spanish

fr: French

Note

Character strings in Code, Label, and Description must all be same length as the number of performance indicators (nPIs) in Value

Slots

Code

A short code for the Performance Indicators for this object. A character string length nPI or a named list for multi-language support. See Details

Label

A short label for the Performance Indicators for this object. Used to label axes on charts. Can be longer than Code but recommended to keep short as possible so it shows clearly in plots and tables. A character string length nPI or a named list for multi-language support. See Details

Description

A description for the Performance Indicators for this object. Can include Markdown, see Examples. A character string length nPI or a named list for multi-language support. See Details

Value

A numeric array with the stochastic performance indicator values for each operating model (OM), management procedure (MP), and performance indicator (PI). Dimensions: c(nOM, nMP, and nPI)

Preset

An optional named list for the preset buttons in the App(). The name of the list element will appear as a button in the App().

Color

A character vector length 2 of colors for the maximum and minimum values in the chart.

MinValue

Numeric vector length nPI with the minimum possible value for the respective PIs. Defaults to minimum PI value in Value (averaged across OMs in some cases)

MaxValue

Numeric vector length nPI with the maximum possible value (i.e., best performance) for the respective PIs. Defaults to maximum PI value in Value (averaged across OMs in some cases).

See also

Examples

# Generate dummy values
nOM <- 2
nMP <- 4
nPI <- 4

values <- array(NA, dim=c(nOM, nMP, nPI))

pi_means <- runif(nPI, 5, 50)
for (om in 1:nOM) {
  for (mp in 1:nMP) {
    for (pi in 1:nPI) {
      values[om, mp, pi] <- rlnorm(1,log(pi_means[pi]), 0.4)
    }
  }
}

# Create and populate Object
quilt <- Quilt(Code=c('PI1', 'PI2', 'PI3', 'PI4'),
               Label=c('Performance Indicator 1',
                       'Performance Indicator 2',
                       'Performance Indicator 3',
                       'Performance Indicator 4'),
               Description = c('This is the description for PI 1',
                               'This is the description for PI 2',
                               'This is the description for PI 3',
                               'This is the description for PI 4'),
               Value=values)

# Check
Check(quilt)
#> 
#> ── Checking: "Quilt" ──
#> 
#> ✔ Complete

# Add to `Slick` object
slick <- Slick()
Quilt(slick) <- quilt

# Plots
plotQuilt(slick)
#> ℹ Note: `MPs` is empty. Using default MP names and colors

plotQuilt(slick, kable=TRUE)
#> ℹ Note: `MPs` is empty. Using default MP names and colors

MP

PI1

PI2

PI3

PI4

MP 1

29

29

14

11.0

MP 2

37

28

20

9.7

MP 3

48

53

27

8.2

MP 4

92

35

18

14.0