Plot Quilt

Create a Quilt plot (unless shading==FALSE in which case it's just a table)

Usage

plotQuilt(
  slick,
  MP_label = "Code",
  minmax = FALSE,
  shading = TRUE,
  kable = FALSE,
  signif = 2,
  alpha = 0.5
)

Arguments

slick

A Slick-class() object

MP_label

Label to use for the MPs. Either Code or Label. Description works as well, but you probably don't want to do that.

minmax

Logical. Color shading from min to max values in each column? If TRUE, ignores MinValue(quilt) and MaxValue(quilt)

shading

Logical. Color shading for the columns?

kable

Logical. Return a kable object?

signif

Numeric Number of significant figures

alpha

Numeric value. Transparency for color shading

Value

A DT::datatable or a knitr::kable object

Details

The columns are color shaded from light (lowest values) to dark (highest values).

Colors are set in Color(quilt).

The color shading has 10 steps, from MinValue(quilt) to MaxValue(quilt) for each Performance Indicator. If those values are missing (NA) for a given PI, colors are shaded from lowest to highest values. If minmax==TRUE, MinValue(quilt) and MaxValue(quilt) are ignored.

See also